requirements.txt
emoji
pymongo
python-telegram-bot
